backend/requirements.txt
virtualenv==20.25.1
fastapi==0.115.12
python-jose==3.3.0
passlib==1.7.4
httpx==0.28.1
uvicorn==0.34.0
gunicorn==23.0.0
pydantic==2.11.2
python-dotenv==1.1.0
minio==7.2.15
python-multipart==0.0.20